# backend/docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
  app.APIResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
    type: object
...
  app.ListNodesResponse:
    properties:
      nodes:
        items:
          $ref: '#/definitions/types.Node'
        type: array
      total:
        type: integer
    type: object
  app.ListNodesWithDiscountResponse:
    properties:
      nodes:
        items:
          $ref: '#/definitions/app.NodesWithDiscount'
        type: array
      total:
        type: integer
    type: object
...
  app.NodeStoragePoolResponse:
    properties:
      pools:
        items:
          $ref: '#/definitions/app.Pool'
        type: array
    type: object
  app.NodesWithDiscount:
    properties:
      certificationType:
        type: string
      city:
        type: string
      country:
        type: string
      cpu_benchmark:
        $ref: '#/definitions/types.CpuBenchmark'
      created:
        type: integer
      dedicated:
        type: boolean
      discount_price:
        type: number
      dmi:
        $ref: '#/definitions/types.Dmi'
      extraFee:
        type: integer
      farm_free_ips:
        type: integer
      farmId:
        type: integer
      farmName:
        type: string
      farmingPolicyId:
        type: integer
      features:
        items:
          type: string
        type: array
      gpus:
        items:
          $ref: '#/definitions/types.NodeGPU'
        type: array
      gridVersion:
        type: integer
      healthy:
        type: boolean
      id:
        type: string
      inDedicatedFarm:
        type: boolean
      location:
        $ref: '#/definitions/types.Location'
      nodeId:
        type: integer
      num_gpu:
        type: integer
      power:
        $ref: '#/definitions/types.NodePower'
      price_usd:
        type: number
      publicConfig:
        $ref: '#/definitions/types.PublicConfig'
      rentContractId:
        type: integer
      rentable:
        type: boolean
      rented:
        type: boolean
      rentedByTwinId:
        type: integer
      serialNumber:
        type: string
      speed:
        $ref: '#/definitions/types.Speed'
      status:
        type: string
      total_resources:
        $ref: '#/definitions/types.Capacity'
      twinId:
        type: integer
      updatedAt:
        type: integer
      uptime:
        type: integer
      used_resources:
        $ref: '#/definitions/types.Capacity'
    type: object
...
  app.Pool:
    properties:
      free:
        description: free space in bytes
        type: integer
      name:
        type: string
      type:
        description: type of the disk wither ssd or hdd
        type: string
    type: object
...
  app.ReserveNodeResponse:
    properties:
      email:
        type: string
      node_id:
        type: integer
      workflow_id:
        type: string
    type: object
...
  app.TwinResponse:
    properties:
      account_id:
        type: string
      public_key:
        type: string
      relay:
        type: string
      twin_id:
        type: integer
    type: object
  app.UnreserveNodeResponse:
    properties:
      contract_id:
        type: integer
      email:
        type: string
      workflow_id:
        type: string
    type: object
...
  gridtypes.Unit:
    enum:
    - 1024
    - 1048576
    - 1073741824
    - 1099511627776
    type: integer
    x-enum-varnames:
    - Kilobyte
    - Megabyte
    - Gigabyte
    - Terabyte
...
  types.BIOS:
    properties:
      vendor:
        type: string
      version:
        type: string
    type: object
  types.Baseboard:
    properties:
      manufacturer:
        type: string
      product_name:
        type: string
    type: object
  types.Capacity:
    properties:
      cru:
        type: integer
      hru:
        $ref: '#/definitions/gridtypes.Unit'
      mru:
        $ref: '#/definitions/gridtypes.Unit'
      sru:
        $ref: '#/definitions/gridtypes.Unit'
    type: object
  types.CpuBenchmark:
    properties:
      multi:
        type: number
      node_twin_id:
        type: integer
      single:
        type: number
      threads:
        type: integer
      updated_at:
        type: integer
      workloads:
        type: integer
    type: object
  types.Dmi:
    properties:
      baseboard:
        $ref: '#/definitions/types.Baseboard'
      bios:
        $ref: '#/definitions/types.BIOS'
      memory:
        items:
          $ref: '#/definitions/types.Memory'
        type: array
      node_twin_id:
        type: integer
      processor:
        items:
          $ref: '#/definitions/types.Processor'
        type: array
      updated_at:
        type: integer
    type: object
  types.Location:
    properties:
      city:
        type: string
      country:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      region:
        type: string
    type: object
  types.Memory:
    properties:
      manufacturer:
        type: string
      type:
        type: string
    type: object
  types.Node:
    properties:
      certificationType:
        type: string
      city:
        type: string
      country:
        type: string
      cpu_benchmark:
        $ref: '#/definitions/types.CpuBenchmark'
      created:
        type: integer
      dedicated:
        type: boolean
      dmi:
        $ref: '#/definitions/types.Dmi'
      extraFee:
        type: integer
      farm_free_ips:
        type: integer
      farmId:
        type: integer
      farmName:
        type: string
      farmingPolicyId:
        type: integer
      features:
        items:
          type: string
        type: array
      gpus:
        items:
          $ref: '#/definitions/types.NodeGPU'
        type: array
      gridVersion:
        type: integer
      healthy:
        type: boolean
      id:
        type: string
      inDedicatedFarm:
        type: boolean
      location:
        $ref: '#/definitions/types.Location'
      nodeId:
        type: integer
      num_gpu:
        type: integer
      power:
        $ref: '#/definitions/types.NodePower'
      price_usd:
        type: number
      publicConfig:
        $ref: '#/definitions/types.PublicConfig'
      rentContractId:
        type: integer
      rentable:
        type: boolean
      rented:
        type: boolean
      rentedByTwinId:
        type: integer
      serialNumber:
        type: string
      speed:
        $ref: '#/definitions/types.Speed'
      status:
        type: string
      total_resources:
        $ref: '#/definitions/types.Capacity'
      twinId:
        type: integer
      updatedAt:
        type: integer
      uptime:
        type: integer
      used_resources:
        $ref: '#/definitions/types.Capacity'
    type: object
  types.NodeGPU:
    properties:
      contract:
        type: integer
      device:
        type: string
      id:
        type: string
      node_twin_id:
        type: integer
      updated_at:
        type: integer
      vendor:
        type: string
      vram:
        type: integer
    type: object
  types.NodePower:
    properties:
      state:
        type: string
      target:
        type: string
    type: object
  types.Processor:
    properties:
      thread_count:
        type: string
      version:
        type: string
    type: object
  types.PublicConfig:
    properties:
      domain:
        type: string
      gw4:
        type: string
      gw6:
        type: string
      ipv4:
        type: string
      ipv6:
        type: string
    type: object
  types.Speed:
    properties:
      download:
        description: in bit/sec
        type: number
      node_twin_id:
        type: integer
      tcp_download_ipv6:
        description: in bit/sec
        type: number
      tcp_upload_ipv6:
        description: in bit/sec
        type: number
      udp_download_ipv4:
        description: in bit/sec
        type: number
      udp_download_ipv6:
        description: in bit/sec
        type: number
      udp_upload_ipv4:
        description: in bit/sec
        type: number
      udp_upload_ipv6:
        description: in bit/sec
        type: number
      updated_at:
        type: integer
      upload:
        description: in bit/sec let's suppose default is ipv4/tcp
        type: number
    type: object
info:
  contact: {}
  description: API documentation for Mycelium Cloud.
  title: Mycelium Cloud API
  version: "1.0"
paths:
...
  /nodes:
    get:
      consumes:
      - application/json
      description: List all nodes from the grid proxy (no user-specific filtering)
      operationId: list-all-grid-nodes
      parameters:
      - description: 'Filter by healthy nodes (default: false)'
        in: query
        name: healthy
        type: boolean
      - description: 'Limit the number of nodes returned (default: 50)'
        in: query
        name: size
        type: integer
      - description: 'page number (default: 1)'
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: All grid nodes retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/app.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/app.ListNodesResponse'
              type: object
        "400":
          description: Invalid filter parameters
          schema:
            $ref: '#/definitions/app.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/app.APIResponse'
      summary: List all grid nodes
      tags:
      - nodes
  /nodes/{node_id}/storage-pool:
    get:
      consumes:
      - application/json
      description: Returns node storage pool
      operationId: get-node-storage-pool
      parameters:
      - description: Node ID
        in: path
        name: node_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Node storage pool is retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/app.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/app.NodeStoragePoolResponse'
              type: object
        "400":
          description: Bad Request or Invalid params
          schema:
            $ref: '#/definitions/app.APIResponse'
        "404":
          description: Node not found
          schema:
            $ref: '#/definitions/app.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.APIResponse'
      summary: Get node storage pool
      tags:
      - nodes
...
  /twins/{twin_id}/account:
    get:
      consumes:
      - application/json
      description: Retrieve the account ID associated with a specific twin ID
      parameters:
      - description: Twin ID
        in: path
        name: twin_id
        required: true
        type: integer
      - description: Pagination limit
        in: query
        name: limit
        type: integer
      - description: Pagination offset
        in: query
        name: offset
        type: integer
      - description: Other optional filter params
        in: query
        name: filterParam
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Account ID is retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/app.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/app.TwinResponse'
              type: object
        "400":
          description: Bad Request or Invalid params
          schema:
            $ref: '#/definitions/app.APIResponse'
        "404":
          description: Twin ID not found
          schema:
            $ref: '#/definitions/app.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.APIResponse'
      summary: Get account ID by twin ID
      tags:
      - twins
...
  /user/nodes:
    get:
      consumes:
      - application/json
      description: List nodes from proxy [rented nodes first + randomized shared nodes]
      operationId: list-nodes
      parameters:
      - description: 'Filter by healthy nodes (default: true)'
        in: query
        name: healthy
        type: boolean
      - description: 'Filter by rentable nodes (default: true)'
        in: query
        name: rentable
        type: boolean
      - description: 'Limit the number of nodes returned (default: 50)'
        in: query
        name: limit
        type: integer
      - description: 'Offset for pagination (default: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Nodes are retrieved successfully
          schema:
            $ref: '#/definitions/app.APIResponse'
        "400":
          description: Invalid filter parameters
          schema:
            $ref: '#/definitions/app.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/app.APIResponse'
      security:
      - UserMiddleware: []
      summary: List nodes
      tags:
      - nodes
  /user/nodes/{node_id}:
    post:
      consumes:
      - application/json
      description: Reserves a node for a user
      operationId: reserve-node
      parameters:
      - description: Node ID
        in: path
        name: node_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Node reservation in progress
          schema:
            allOf:
            - $ref: '#/definitions/app.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/app.ReserveNodeResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/app.APIResponse'
        "404":
          description: No nodes are available for rent.
          schema:
            $ref: '#/definitions/app.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.APIResponse'
      security:
      - UserMiddleware: []
      summary: Reserve node
      tags:
      - nodes
  /user/nodes/rentable:
    get:
      consumes:
      - application/json
      description: Retrieves a list of rentable nodes from the grid proxy. These are
        healthy nodes that are available for rent.
      operationId: list-rentable-nodes
      produces:
      - application/json
      responses:
        "200":
          description: Rentable nodes retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/app.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/app.ListNodesWithDiscountResponse'
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/app.APIResponse'
      summary: List rentable nodes
      tags:
      - nodes
  /user/nodes/rented:
    get:
      consumes:
      - application/json
      description: Returns a list of reserved nodes for a user
      operationId: list-reserved-nodes
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/app.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/app.ListNodesWithDiscountResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.APIResponse'
      security:
      - UserMiddleware: []
      summary: List reserved nodes
      tags:
      - nodes
  /user/nodes/unreserve/{contract_id}:
    delete:
      consumes:
      - application/json
      description: Unreserve a node for a user
      operationId: unreserve-node
      parameters:
      - description: Contract ID
        in: path
        name: contract_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            allOf:
            - $ref: '#/definitions/app.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/app.UnreserveNodeResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/app.APIResponse'
        "404":
          description: User is not found
          schema:
            $ref: '#/definitions/app.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.APIResponse'
      security:
      - UserMiddleware: []
      summary: Unreserve node
      tags:
      - nodes
...
swagger: "2.0"